refactor(test): replace deprecated extension functions with MockMvcRequestBuilders
- Replace get/post/put extension functions with MockMvcRequestBuilders.get/post/put
